3. Quick sort

+++ sort.h
@@ -35,7 +35,7 @@ void selection_sort(int *array, size_t size);
 
 /* 3. Quick sort  */
 void swap_ints(int *a, int *b);
-int lomuto_partition(int *array, ssize_t low, ssize_t high, size_t size);
+size_t lomuto_partition(int *array, ssize_t low, ssize_t high, size_t size);
 void quicksort_recursive(int *array, ssize_t low, ssize_t high, size_t size);
 void quick_sort(int *array, size_t size);
